w3c-reports/12092014/mac/runned-tests/XMLHttpRequest/getresponseheader-cookies-and-more-expected.txt

CONSOLE MESSAGE: Refused to get unsafe header "set-cookie"
CONSOLE MESSAGE: Refused to get unsafe header "set-cookie2"
CONSOLE MESSAGE: Refused to get unsafe header "set-cookie"
CONSOLE MESSAGE: Refused to get unsafe header "set-cookie2"
CONSOLE MESSAGE: Refused to get unsafe header "set-cookie"
CONSOLE MESSAGE: Refused to get unsafe header "set-cookie2"

PASS XMLHttpRequest: getResponseHeader() custom/non-existent headers and cookies 
